requirements.txt
Jinja2==2.9.6
MarkupSafe==1.0
PyYAML==3.12
wheel==0.24.0
termcolor==1.1.0
fire==0.1.0

-e .
